write documentation

+++ jcdb-core/src/main/kotlin/org/utbot/jcdb/impl/features/UsageRequest.kt
@@ -14,6 +14,6 @@ data class UsageFeatureRequest(
 
 class UsageFeatureResponse(
     val source: ClassSource,
-    val offsets: ByteArray
+    val offsets: ShortArray
 )
